requirements.txt
alembic
aniso8601
astroid==2.4.2
async-generator
attrs
autopep8
bcrypt
beautifulsoup4
blinker
bs4
certifi
cffi
chardet
click
colorama
configparser
coverage
crayons
cryptography
cycler
decorator
dnspython
email-validator
Flask
Flask-Bcrypt
Flask-Cors
Flask-Dance
Flask-DebugToolbar
Flask-Login
Flask-Mail
Flask-Migrate
Flask-MySQLdb
Flask-OAuth
Flask-OAuthlib
Flask-RESTful
Flask-Script
Flask-SQLAlchemy
Flask-Testing
Flask-WTF
greenlet
gunicorn
h11
httplib2
idna
isort
itsdangerous
Jinja2
kiwisolver
lazy
lazy-object-proxy
lxml
Mako
MarkupSafe
matplotlib
mccabe
mysql
mysql-connector-python
mysqlclient
networkx
oauth2
oauthlib
outcome
Pillow
protobuf
pycodestyle
pycparser
pylint
PyMySQL
pyOpenSSL
pyparsing
python-dateutil
python-editor
pytz
requests
requests-oauthlib
scipy
selenium
six
sniffio
sortedcontainers
soupsieve
SQLAlchemy
SQLAlchemy-Utils
toml
trio
trio-websocket
typed-ast
urllib3
URLObject
webdriver-manager
Werkzeug
wincertstore
wrapt
wsproto
WTForms
xlrd
numpy
pandas
